+++ wrapper/screen/test_check.go
@@ -20,7 +20,7 @@ func NewTestCheck() screen.Screen {
 		),
 		*text.NewLine(
 			"-",
-			style.SpecFill(uint(sizeTitle)),
+			style.SpecFill(sizeTitle),
 		),
 		*text.EmptyLine(),
 	}
